Added a quick fix for the tone curve filter by Marcelo Barbudas.

+++ framework/Source/GPUImageToneCurveFilter.m
@@ -281,7 +281,7 @@ NSString *const kGPUImageToneCurveFragmentShaderString = SHADER_STRING
         CGPoint cur = [[points objectAtIndex:i] CGPointValue];
         CGPoint next = [[points objectAtIndex:(i+1)] CGPointValue];
         
-        for(int x=cur.x;x<next.x;x++) 
+        for(int x=cur.x;x<(int)next.x;x++) 
         {
             double t = (double)(x-cur.x)/(next.x-cur.x);
